# ShaderCodeCursor. Optimization of lines parsing