--- a/packages/dev/core/src/Engines/Processors/shaderCodeCursor.ts
+++ b/packages/dev/core/src/Engines/Processors/shaderCodeCursor.ts
@@ -15,6 +15,11 @@ export class ShaderCodeCursor {
         this._lines.length = 0;
 
         for (const line of value) {
+            // Skip empty lines
+            if (!line || line === "\r") {
+                continue;
+            }
+
             // Prevent removing line break in macros.
             if (line[0] === "#") {
                 this._lines.push(line);
@@ -22,22 +27,45 @@ export class ShaderCodeCursor {
             }
 
             // Do not split single line comments
-            if (line.trim().startsWith("//")) {
+            const trimmedLine = line.trim();
+
+            if (!trimmedLine) {
+                continue;
+            }
+
+            if (trimmedLine.startsWith("//")) {
                 this._lines.push(line);
                 continue;
             }
 
-            const split = line.split(";");
+            // Work with semicolon in the line
+            const semicolonIndex = line.indexOf(";");
 
-            for (let index = 0; index < split.length; index++) {
-                let subLine = split[index];
-                subLine = subLine.trim();
+            if (semicolonIndex === -1) {
+                // No semicolon in the line
+                this._lines.push(trimmedLine);
+            } else if (semicolonIndex === line.length - 1 || semicolonIndex === trimmedLine.length - 1) {
+                // Semicolon at the end of the line
+                this._lines.push(line);
+            } else {
+                // Semicolon in the middle of the line
+                const split = line.split(";");
 
-                if (!subLine) {
-                    continue;
+                for (let index = 0; index < split.length; index++) {
+                    let subLine = split[index];
+
+                    if (!subLine) {
+                        continue;
+                    }
+
+                    subLine = subLine.trim();
+
+                    if (!subLine) {
+                        continue;
+                    }
+
+                    this._lines.push(subLine + (index !== split.length - 1 ? ";" : ""));
                 }
-
-                this._lines.push(subLine + (index !== split.length - 1 ? ";" : ""));
             }
         }
     }