@bjsplat 2023-06-02
Snapshot stored with reference name:
refs/pull/13935/merge

Test environment: 
https://babylonsnapshots.z22.web.core.windows.net/refs/pull/13935/merge/index.html

To test a playground add it to the URL, for example:

https://babylonsnapshots.z22.web.core.windows.net/refs/pull/13935/merge/index.html#WGZLGJ#4600

Links to test babylon tools with this snapshot:

https://playground.babylonjs.com/?snapshot=refs/pull/13935/merge
https://sandbox.babylonjs.com/?snapshot=refs/pull/13935/merge
https://gui.babylonjs.com/?snapshot=refs/pull/13935/merge
https://nme.babylonjs.com/?snapshot=refs/pull/13935/merge

To test the snapshot in the playground with a playground ID add it after the snapshot query string:

https://playground.babylonjs.com/?snapshot=refs/pull/13935/merge#BCU1XR#0