@RaananW 2023-06-05
We had the same discussion in a different PR. Our UMD is compiled this way due to back-compat. It will be the best to use a different type of iteration.

But is it in this PR? it seems like this PR builds correctly.