@RaananW 2023-06-08
> > There is a shader compilation issue in the webgl2 visualization tests
> 
> @RaananW as I can see in logs, there is an error with npm install:
> https://dev.azure.com/babylonjs/ContinousIntegration/_build/results?buildId=25611&view=logs&j=1f876048-d530-54e9-d49b-d5fedfaadf19&t=6aa1df57-0aab-5db6-f94c-8054e58bee0c
> 
> ```
> /usr/bin/bash --noprofile --norc /home/vsts/work/_temp/73c5cb84-7ea8-4efc-8191-59f6177859ff.sh
> npm ERR! code 1
> npm ERR! path /home/vsts/work/1/s/node_modules/puppeteer
> npm ERR! command failed
> npm ERR! command sh -c -- node install.js
> npm ERR! Requesting latest Firefox Nightly version from https://product-details.mozilla.org/1.0/firefox_versions.json
> npm ERR! ERROR: Failed to set up Firefox Nightly r116.0a1! Set "PUPPETEER_SKIP_DOWNLOAD" env variable to skip download.
> npm ERR! Error: Download failed: server returned code 404. URL: https://archive.mozilla.org/pub/firefox/nightly/latest-mozilla-central/firefox-116.0a1.en-US.linux-x86_64.tar.bz2
> npm ERR!     at /home/vsts/work/1/s/node_modules/puppeteer/lib/cjs/puppeteer/node/BrowserFetcher.js:371:27
> npm ERR!     at ClientRequest.requestCallback (/home/vsts/work/1/s/node_modules/puppeteer/lib/cjs/puppeteer/node/BrowserFetcher.js:500:13)
> npm ERR!     at Object.onceWrapper (node:events:628:26)
> npm ERR!     at ClientRequest.emit (node:events:513:28)
> npm ERR!     at HTTPParser.parserOnIncomingClient (node:_http_client:693:27)
> npm ERR!     at HTTPParser.parserOnHeadersComplete (node:_http_common:128:17)
> npm ERR!     at TLSSocket.socketOnData (node:_http_client:534:22)
> npm ERR!     at TLSSocket.emit (node:events:513:28)
> npm ERR!     at addChunk (node:internal/streams/readable:315:12)
> npm ERR!     at readableAddChunk (node:internal/streams/readable:289:9)
> ```
> 
> Is this related to my updates?

Feels like a temporary npm issue. I have triggered it again, let's see if it fails with the same problem again