@bjsplat 2023-06-08
Visualization tests for webgl2 have failed. If some tests failed because the snapshots do not match, the report can be found at

https://babylonsnapshots.z22.web.core.windows.net/refs/pull/13935/merge/testResults/webgl2/index.html

If tests were successful afterwards, this report might not be available anymore.