@Dok11 2023-06-11
@RaananW could you help me to find what is this test?

> ![image](https://github.com/BabylonJS/Babylon.js/assets/2697890/580659b3-006e-4ff4-acab-31081810b1be)

Searcing "Glossiness" by sources does not give image similar to

> ![image](https://github.com/BabylonJS/Babylon.js/assets/2697890/3f97ceaf-b3ee-4b32-a243-dbb237b38ccb)

Most similar is 

> ![image](https://github.com/BabylonJS/Babylon.js/assets/2697890/82cc7f80-d3ed-49a9-9c16-7373fd98ec7d)

from PBRSpecularGlossinessMaterial.png

I trying to find the playground with bugs where I can check this PR changes.